refactor: move API logic to separate class library
