+++ TwitchApi/TwitchAPi.Client/Data/TwitchClient.cs
@@ -1,4 +1,4 @@
-﻿namespace TwitchApi.Twitch.Data;
+﻿namespace TwitchAPi.Client.Data;
 public record ClientTwitchAuth(
     string ClientId,
     string ClientSecret
