+++ TwitchApi/TwitchAPi.Client/Data/UserTwitchAuth.cs
@@ -1,4 +1,4 @@
-﻿namespace TwitchApi.Twitch.Data;
+﻿namespace TwitchAPi.Client.Data;
 
 public record UserTwitchAuth(
     string? AccessToken,
