+++ TwitchApi/TwitchAPi.Client/Responses/TokenResponse.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace TwitchApi.Twitch.Responses;
+namespace TwitchAPi.Client.Responses;
 
 public record TokenResponse(
     [property: JsonPropertyName("access_token")]
