--- TwitchApi/TwitchApi/Extensions.cs
@@ -1,23 +1,0 @@
-﻿using System.Collections.Specialized;
-using System.Web;
-
-namespace TwitchApi
-{
-    public static class Extensions
-    {
-        public static string ToQueryString(this NameValueCollection nvc)
-        {
-            if (nvc == null || nvc.Count == 0)
-                return string.Empty;
-
-            var array = (
-                from key in nvc.AllKeys
-                where key != null
-                from value in nvc.GetValues(key) ?? Array.Empty<string>()
-                select $"{HttpUtility.UrlEncode(key)}={HttpUtility.UrlEncode(value)}"
-            ).ToArray();
-
-            return array.Length > 0 ? "?" + string.Join("&", array) : string.Empty;
-        }
-    }
-}
